refactor: use type members in place of generics
